Extract command line parser to utility
