+++ src/main/java/HttpServer/Server.java
@@ -1,9 +1,8 @@
 package HttpServer;
 
+import HttpServer.utility.CommandLineParser;
 import HttpServer.utility.Logger;
 import HttpServer.utility.VerboseLogger;
-
-import java.util.Objects;
 
 public class Server {
     public static void main(String[] args) throws Exception {
@@ -20,39 +19,4 @@ public class Server {
         listener.start();
     }
 
-    // Parsing command line arguments when the server is started.
-    private static class CommandLineParser {
-        private final Logger logger;
-        private String[] args;
-        private int port;
-        private String directory;
-
-        public CommandLineParser(String[] args, int port, String directory, Logger logger) {
-            this.args = args;
-            this.port = port;
-            this.directory = directory;
-            this.logger = logger;
-        }
-
-        public int getPort() {
-            return port;
-        }
-
-        public String getDirectory() {
-            return directory;
-        }
-
-        public CommandLineParser invoke() {
-            logger.log("Processing command line arguments");
-            for (int i=0; i < args.length; i++) {
-                String token = args[i];
-                if (Objects.equals(token, "-p")) {
-                    port = Integer.parseInt(args[i+1]);
-                } else if (token.equals("-d")) {
-                    directory = args[i+1];
-                }
-            }
-            return this;
-        }
-    }
 }
